+++ docs/background.md
@@ -8,19 +8,19 @@ Signature: `BackgroundTask(func, *args, **kwargs)
 ```python
 from starlette.applications import Starlette
 from starlette.responses import JSONResponse
-from starlette.tasks import BackgroundTask
+from starlette.background import BackgroundTask
 
 app = Starlette()
 
 @app.route('/start_task/{arg}')
 def init_task(request, arg):
-    task = BackgroundTask(some_long_task, arg)
+    task = BackgroundTask(some_task, arg)
     return JSONResponse(
         'Task was initiated',
         202,
         background=task
     )
 
-def some_long_task(arg):
+async def some_task(arg):
     pass
 ```
